added a better flow for search

+++ app/src/main/java/com/dvt/weatherapp/domain/models/WeatherList.kt
@@ -12,7 +12,7 @@ data class WeatherList (
     @SerializedName("main")
     var main: MainForecast?,
     @SerializedName("pop")
-    var pop: Int?,
+    var pop: Double?,
     @SerializedName("sys")
     var sys: SysForecast?,
     @SerializedName("visibility")
